do not use random coordinates as a hack to store cleaning progress (#53)
Make this more JobDriver_CleanFilth-like (where this was presumably copied
from). Since RimWorld doesn't make it easy to add extra data to its classes,
make an extra Dictionary-based class that stores the JobDriver-related data,
and saving/loading of the data is done as part of the relevant JobDriver
ExposeData().

This not strictly backwards compatible with previous saves in case a save
is done while the cleaning is in progress, but the code handles that case
fine (progress is reset and the infinity resulting from the division by zero
makes progressbar full until the cleaning finishes).